+++ app/Http/Controllers/ProductController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\ProductEditRequest;
 use App\Http\Requests\ProductRequest;
 use Illuminate\Http\Request;
 use App\Models\Product;
@@ -15,9 +16,16 @@ class ProductController extends Controller
         ]);
     }
     public function index($id){
-        return response()->json([
-            'content' => Product::find($id)
-        ]);
+        $product = Product::find($id);
+        if(!$product){
+            return response()->json([
+                'message' => 'Товар не существует',
+            ],404);
+        } else {
+            return response()->json([
+                'content' => $product
+            ]);
+        }
     }
     public function store(ProductRequest $request){
         $product = Product::create($request->all());
@@ -29,11 +37,30 @@ class ProductController extends Controller
             ]
         ]);
     }
-    public function update(Request $request){
+    public function update($id,ProductEditRequest $request){
+        $product = Product::find($id);
+        if(!$product){
+            return response()->json([
+                'message' => 'Товара не существует'
+            ]);
+        }
+        $product->update($request->all());
+
+        return response()->json([
+            'content' => [
+                'id' => $product->id,
+                'message' => 'Данные обновлены'
+            ]
+        ], 200);
 
     }
     public function destroy($id){
         $product = Product::find($id);
+        if(!$product){
+            return response()->json([
+                'message' => 'Товара не существует',
+            ],404);
+        }
         $product->delete();
         return response()->json([
             'content' => [
